Fixes #1938
Split pbench-index.py into modules
collect_tb() - collects the TB that needs to be indexed
process_tb() - process for indexing
