+++ server/bin/gold/test-7.26.txt
@@ -1134,7 +1134,7 @@ drwxrwxr-x          - logs/pbench-audit-server
 -rw-rw-r--          0 logs/pbench-audit-server/pbench-audit-server.error
 -rw-rw-r--        437 logs/pbench-audit-server/pbench-audit-server.log
 drwxrwxr-x          - logs/pbench-index-re
--rw-rw-r--       3882 logs/pbench-index-re/pbench-index-re.log
+-rw-rw-r--       4025 logs/pbench-index-re/pbench-index-re.log
 drwxrwxr-x          - pbench-move-results-receive
 drwxrwxr-x          - pbench-move-results-receive/fs-version-002
 drwxrwxr-x          - quarantine
@@ -1189,16 +1189,16 @@ drwxrwxr-x          - tmp
 ----- pbench-audit-server/pbench-audit-server.log
 +++++ pbench-index-re/pbench-index-re.log
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- pbench-index-re.run-1970-01-01T00:00:42-UTC: starting
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- Preparing to index 1 tar balls
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- update_templates [start]
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- Preparing to index 1 tar balls
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- update_templates [start]
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer update_templates -- done templates (start ts: 1970-01-01T00:00:42-UTC, end ts: 1970-01-01T00:00:42-UTC, duration: 0.00s, successes: 12, retries: 0)
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- update_templates [end]
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- update_templates [end]
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.report post_status -- posted status (start ts: 1970-01-01T00:00:42-UTC, end ts: 1970-01-01T00:00:42-UTC, duration: 0.00s, successes: 1, duplicates: 0, failures: 0, retries: 0)
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- start processing list of tar balls
-1970-01-01T00:00:42.000000 INFO pbench-index-re.pbench-index main -- Starting /var/tmp/pbench-test-server/test-7.26/pbench/archive/fs-version-001/ctlrA/TO-RE-INDEX/linpack_mock_2020.02.28T19.10.55.tar.xz (size 20848)
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- open tar ball
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- generator setup
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- begin indexing
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- start processing list of tar balls
+1970-01-01T00:00:42.000000 INFO pbench-index-re.indexing_tarballs process_tb -- Starting /var/tmp/pbench-test-server/test-7.26/pbench/archive/fs-version-001/ctlrA/TO-RE-INDEX/linpack_mock_2020.02.28T19.10.55.tar.xz (size 20848)
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- open tar ball
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- generator setup
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- begin indexing
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer make_all_actions -- start
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer mk_run_action -- start
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer mk_sosreports -- start
@@ -1211,12 +1211,12 @@ drwxrwxr-x          - tmp
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer mk_result_data_actions -- start
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer mk_result_data_actions -- end [6 result documents]
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexer make_all_actions -- end
-1970-01-01T00:00:42.000000 INFO pbench-index-re.pbench-index main -- done indexing (start ts: 1970-01-01T00:00:42-UTC, end ts: 1970-01-01T00:00:42-UTC, duration: 0.00s, successes: 26, duplicates: 0, failures: 0, retries: 0)
-1970-01-01T00:00:42.000000 INFO pbench-index-re.pbench-index main -- run-1970-01-01T00:00:42-UTC: ctlrA/linpack_mock_2020.02.28T19.10.55.tar.xz: success
-1970-01-01T00:00:42.000000 INFO pbench-index-re.pbench-index main -- Finished /var/tmp/pbench-test-server/test-7.26/pbench/archive/fs-version-001/ctlrA/TO-RE-INDEX/linpack_mock_2020.02.28T19.10.55.tar.xz (size 20848)
+1970-01-01T00:00:42.000000 INFO pbench-index-re.indexing_tarballs process_tb -- done indexing (start ts: 1970-01-01T00:00:42-UTC, end ts: 1970-01-01T00:00:42-UTC, duration: 0.00s, successes: 26, duplicates: 0, failures: 0, retries: 0)
+1970-01-01T00:00:42.000000 INFO pbench-index-re.indexing_tarballs process_tb -- run-1970-01-01T00:00:42-UTC: ctlrA/linpack_mock_2020.02.28T19.10.55.tar.xz: success
+1970-01-01T00:00:42.000000 INFO pbench-index-re.indexing_tarballs process_tb -- Finished /var/tmp/pbench-test-server/test-7.26/pbench/archive/fs-version-001/ctlrA/TO-RE-INDEX/linpack_mock_2020.02.28T19.10.55.tar.xz (size 20848)
 1970-01-01T00:00:42.000000 WARNING pbench-index-re.indexer dump_opctx -- ** Errors encountered while indexing: [{"counters": {"sample_missing_timeseries": 6}, "object": "ResultData", "tbname": "/var/tmp/pbench-test-server/test-7.26/pbench/archive/fs-version-001/ctlrA/linpack_mock_2020.02.28T19.10.55.tar.xz"}]
-1970-01-01T00:00:42.000000 DEBUG pbench-index-re.pbench-index main -- stopped processing list of tar balls
-1970-01-01T00:00:42.000000 INFO pbench-index-re.pbench-index main -- pbench-index-re.run-1970-01-01T00:00:42-UTC: indexed 1 (skipped 0) results, 0 errors
+1970-01-01T00:00:42.000000 DEBUG pbench-index-re.indexing_tarballs process_tb -- stopped processing list of tar balls
+1970-01-01T00:00:42.000000 INFO pbench-index-re.indexing_tarballs process_tb -- pbench-index-re.run-1970-01-01T00:00:42-UTC: indexed 1 (skipped 0) results, 0 errors
 1970-01-01T00:00:42.000000 DEBUG pbench-index-re.report post_status -- posted status (start ts: 1970-01-01T00:00:42-UTC, end ts: 1970-01-01T00:00:42-UTC, duration: 0.00s, successes: 1, duplicates: 0, failures: 0, retries: 0)
 ----- pbench-index-re/pbench-index-re.log
 ---- pbench-local/logs
